requirements.txt
Django==5.2
crispy-bootstrap5==2025.4
django-crispy-forms==2.3
pillow==11.1.0
whitenoise==6.9.0